requirements.txt
ics==0.7.2
python-telegram-bot==20.5
PyYAML==5.4.1
PyYAML==6.0.1
